requirements.txt
mcp>=1.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
